+++ applications/play/app-redux/PLAN.txt
