+++ applications/play/app-redux/getInitialState.ts
