+++ applications/play/app-redux/reducer.ts
