+++ applications/play/pages/_app.tsx
@@ -3,7 +3,7 @@ import App, { Container } from "next/app";
 import withRedux from "next-redux-wrapper";
 import { Provider } from "react-redux";
 
-import { createStore } from "../redux";
+import { createStore } from "../app-redux";
 
 /**
  * @param {object} initialState
