+++ applications/play/pages/index.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { connect } from "react-redux";
 
 import Main from "../components/Main";
-import { actions } from "../redux";
+import { actions } from "../app-redux";
 
 function detectPlatform(req) {
   if (req && req.headers) {
